fix: issue with state solved

+++ app/src/main/java/com/manuel28g/carsales/covidworlddata/model/CovidInfoStatus.kt
@@ -24,5 +24,5 @@ data class CovidInfoStatus(
     @SerializedName("active_diff")
     var activeDiff:Long,
     @SerializedName("fatality_rate")
-    var fatalityRate:Long
+    var fatalityRate:Double
 )
